// src/types/product.ts
// types/product.ts

export interface Product {
  id: number
  title: string
  description: string
  category: string
  price: number
  discountPercentage: number
  rating: number
  stock: number
  tags: string[]
  brand: string
  sku: string
  weight: number
  dimensions: {
    width: number
    height: number
    depth: number
  }
  warrantyInformation: string
  shippingInformation: string
  availabilityStatus: string
  reviews: {
    rating: number
    comment: string
    date: string // ISO date string
    reviewerName: string
    reviewerEmail: string
  }[]
  returnPolicy: string
  minimumOrderQuantity: number
  meta: {
    createdAt: string // ISO date string
    updatedAt: string // ISO date string
    barcode: string
    qrCode: string
  }
  images: string[]
  thumbnail: string
}

export interface ProductResponse {
  products: Product[]
  total: number
  skip: number
  limit: number
}